requirements.txt
datasets
sentence-transformers
faiss-cpu
transformers
ollama
langchain 
langchain-huggingface
streamlit
